+++ packages/fluentui/react/src/components/Avatar/Avatar.tsx
@@ -31,6 +31,9 @@ export interface AvatarProps extends UIComponentProps {
   /** The name used for displaying the initials of the avatar if the image is not provided. */
   name?: string;
 
+  /** The avatar can have a square shape. */
+  square?: boolean;
+
   /** Size multiplier. */
   size?: SizeValue;
 
@@ -41,12 +44,14 @@ export interface AvatarProps extends UIComponentProps {
   getInitials?: (name: string) => string;
 }
 
+export type AvatarStylesProps = Pick<AvatarProps, 'size' | 'square'>;
+
 const Avatar: React.FC<WithAsProp<AvatarProps>> & FluentComponentStaticProps<AvatarProps> = props => {
   const context: ProviderContextPrepared = React.useContext(ThemeContext);
   const { setStart, setEnd } = useTelemetry(Avatar.displayName, context.telemetry);
   setStart();
 
-  const { accessibility, className, design, getInitials, label, icon, image, name, size, status, styles, variables } = props;
+  const { accessibility, className, design, getInitials, label, icon, image, name, square, size, status, styles, variables } = props;
 
   const getA11Props = useAccessibility(accessibility, {
     debugName: Avatar.displayName,
@@ -54,7 +59,7 @@ const Avatar: React.FC<WithAsProp<AvatarProps>> & FluentComponentStaticProps<Ava
   });
   const { classes, styles: resolvedStyles } = useStyles(Avatar.displayName, {
     className: Avatar.className,
-    mapPropsToStyles: () => ({ size }),
+    mapPropsToStyles: () => ({ size, square }),
     mapPropsToInlineStyles: () => ({
       className,
       design,
@@ -72,7 +77,7 @@ const Avatar: React.FC<WithAsProp<AvatarProps>> & FluentComponentStaticProps<Ava
         defaultProps: () =>
           getA11Props('image', {
             fluid: true,
-            avatar: true,
+            avatar: !square,
             title: name,
             styles: resolvedStyles.image
           })
@@ -82,7 +87,7 @@ const Avatar: React.FC<WithAsProp<AvatarProps>> & FluentComponentStaticProps<Ava
           defaultProps: () =>
             getA11Props('icon', {
               size,
-              circular: true,
+              circular: !square,
               title: name,
               styles: resolvedStyles.icon
             })
@@ -93,7 +98,7 @@ const Avatar: React.FC<WithAsProp<AvatarProps>> & FluentComponentStaticProps<Ava
           defaultProps: () =>
             getA11Props('label', {
               content: getInitials(name),
-              circular: true,
+              circular: !square,
               title: name,
               styles: resolvedStyles.label
             })
@@ -150,6 +155,7 @@ Avatar.propTypes = {
   icon: customPropTypes.itemShorthandWithoutJSX,
   image: customPropTypes.itemShorthandWithoutJSX,
   label: customPropTypes.itemShorthand,
+  square: PropTypes.bool,
   size: customPropTypes.size,
   status: customPropTypes.itemShorthand,
   getInitials: PropTypes.func