+++ packages/fluentui/react/src/themes/teams/components/Avatar/avatarVariables.ts
@@ -1,6 +1,9 @@
+import { pxToRem } from '../../../../utils';
+
 export interface AvatarVariables {
   avatarBorderColor: string;
   avatarBorderWidth: string;
+  squareAvatarBorderRadius: string;
   statusBorderColor: string;
   statusBorderWidth: string;
 }
@@ -8,6 +11,7 @@ export interface AvatarVariables {
 export default (siteVariables): AvatarVariables => ({
   avatarBorderColor: '',
   avatarBorderWidth: '0',
+  squareAvatarBorderRadius: pxToRem(3),
   statusBorderColor: siteVariables.bodyBackground,
   statusBorderWidth: '2px'
 });